Refactor client.py and models.py into module dirs.
* This make makes room for addition of MLModel functions and classes.
