+++ radiant_mlhub/models/__init__.py
@@ -1,0 +1,6 @@
+"""Extensions of the `PySTAC <https://pystac.readthedocs.io/en/latest/>`_ classes that provide convenience methods for interacting
+with the `Radiant MLHub API <https://docs.mlhub.earth/#radiant-mlhub-api>`_."""
+
+from .Collection import Collection
+from .Dataset import Dataset
+from .MLModel import MLModel
